fix(gates): split the #9637 census rows and deep-equal helpers out for the file cap
