+++ crates/perry-runtime/src/gc/roots/stack_maps.rs
@@ -1997,46 +1997,5 @@ mod decode_tests;
 #[path = "stack_maps_walker_agreement.rs"]
 mod walker_agreement;
 
-/// `PERRY_GC_CENSUS`: bytes owned by the published stack-map index (the
-/// `sections` are `&'static` image bytes and are reported separately as
-/// file-backed).
-pub(crate) fn stack_map_index_census() -> Vec<crate::gc::census::SideTableRow> {
-    use crate::gc::census::vec_bytes;
-    let mut rows = Vec::new();
-    let Some(lock) = STACK_MAPS.published.get() else {
-        return rows;
-    };
-    let Ok(g) = lock.read() else {
-        return rows;
-    };
-    let ix = &g.index;
-    rows.push((
-        "stackmap.sections(file-backed)",
-        ix.sections.len(),
-        ix.sections.iter().map(|s| s.len()).sum(),
-    ));
-    rows.push((
-        "stackmap.functions",
-        ix.functions.len(),
-        vec_bytes(&ix.functions),
-    ));
-    if let Some(e) = ix.eager.as_ref() {
-        rows.push((
-            "stackmap.eager.records",
-            e.records.len(),
-            vec_bytes(&e.records),
-        ));
-        rows.push(("stackmap.eager.roots", e.roots.len(), vec_bytes(&e.roots)));
-        rows.push((
-            "stackmap.eager.derived",
-            e.derived.len(),
-            vec_bytes(&e.derived),
-        ));
-        rows.push((
-            "stackmap.eager.function_starts",
-            e.function_starts.len(),
-            vec_bytes(&e.function_starts),
-        ));
-    }
-    rows
-}
+mod census_rows;
+pub(crate) use census_rows::stack_map_index_census;
